Finished RippleCursor
As of now what I am noticing from initial testing is that the gifs that are played with the cursor perform differently depending on the quality of the video being displayed. This has resulted in me keeping two different cursormodes for the Ripple cursor to deal with low and high performance.